# demo_package/docs/api_specification.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AITM - AI-Powered Threat Modeler",
    "description": "Automated threat modeling using AI agents and MITRE ATT&CK framework",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/analytics/dashboard": {
      "get": {
        "tags": [
          "analytics",
          "Analytics"
        ],
        "summary": "Get comprehensive dashboard metrics",
        "description": "Retrieve executive dashboard metrics including project status, risk assessment, threat landscape, and performance indicators",
        "operationId": "get_dashboard_metrics_api_v1_analytics_dashboard_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "period_days",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 365,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Analysis period in days",
              "default": 30,
              "title": "Period Days"
            },
            "description": "Analysis period in days"
          },
          {
            "name": "include_trends",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Include trend analysis",
              "default": true,
              "title": "Include Trends"
            },
            "description": "Include trend analysis"
          },
          {
            "name": "include_predictions",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Include risk predictions",
              "default": false,
              "title": "Include Predictions"
            },
            "description": "Include risk predictions"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardMetricsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/project/{project_id}": {
      "get": {
        "tags": [
          "analytics",
          "Analytics"
        ],
        "summary": "Get detailed project analytics",
        "description": "Retrieve comprehensive analytics for a specific project including risk assessment, attack paths, and recommendations",
        "operationId": "get_project_analytics_api_v1_analytics_project__project_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          },
          {
            "name": "include_predictions",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Include risk predictions",
              "default": true,
              "title": "Include Predictions"
            },
            "description": "Include risk predictions"
          },
          {
            "name": "include_recommendations",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Include recommendations analysis",
              "default": true,
              "title": "Include Recommendations"
            },
            "description": "Include recommendations analysis"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectAnalyticsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/trends": {
      "get": {
        "tags": [
          "analytics",
          "Analytics"
        ],
        "summary": "Get trend analysis",
        "description": "Retrieve trend analysis for specified metrics over configurable time periods",
        "operationId": "get_trend_analysis_api_v1_analytics_trends_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AnalyticsPeriod",
              "description": "Analysis period",
              "default": "30d"
            },
            "description": "Analysis period"
          },
          {
            "name": "metrics",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "Metrics to analyze",
              "default": [
                "risk_score",
                "project_count"
              ],
              "title": "Metrics"
            },
            "description": "Metrics to analyze"
          },
          {
            "name": "granularity",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(daily|weekly|monthly)$",
              "description": "Data granularity",
              "default": "daily",
              "title": "Granularity"
            },
            "description": "Data granularity"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__models__analytics__TrendAnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/reports/executive": {
      "post": {
        "tags": [
          "analytics",
          "Analytics"
        ],
        "summary": "Generate executive report",
        "description": "Generate comprehensive executive reports with strategic insights and industry comparisons",
        "operationId": "generate_executive_report_api_v1_analytics_reports_executive_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutiveReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutiveReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/analytics/predictions/risk": {
      "post": {
        "tags": [
          "analytics",
          "Analytics"
        ],
        "summary": "Get AI risk predictions",
        "description": "Generate AI-powered risk predictions for projects with configurable prediction horizons",
        "operationId": "get_risk_predictions_api_v1_analytics_predictions_risk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskPredictionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__models__analytics__RiskPredictionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/analytics/export/data": {
      "get": {
        "tags": [
          "analytics",
          "Analytics"
        ],
        "summary": "Export analytics data",
        "description": "Export analytics data in various formats (JSON, CSV)",
        "operationId": "export_analytics_data_api_v1_analytics_export_data_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(json|csv)$",
              "description": "Export format",
              "default": "json",
              "title": "Format"
            },
            "description": "Export format"
          },
          {
            "name": "period_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "description": "Data period",
              "default": 30,
              "title": "Period Days"
            },
            "description": "Data period"
          },
          {
            "name": "include_raw_data",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include raw data points",
              "default": false,
              "title": "Include Raw Data"
            },
            "description": "Include raw data points"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalyticsPeriod": {
        "type": "string",
        "enum": [
          "7d",
          "30d",
          "90d",
          "1y"
        ],
        "title": "AnalyticsPeriod",
        "description": "Time periods for analytics"
      },
...
      "DashboardMetricsResponse": {
        "properties": {
          "period": {
            "type": "string",
            "title": "Period",
            "description": "Analysis period description"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At",
            "description": "Report generation timestamp"
          },
          "project_metrics": {
            "$ref": "#/components/schemas/ProjectMetrics"
          },
          "risk_metrics": {
            "$ref": "#/components/schemas/RiskMetrics"
          },
          "threat_metrics": {
            "$ref": "#/components/schemas/ThreatMetrics"
          },
          "performance_metrics": {
            "$ref": "#/components/schemas/PerformanceMetrics"
          },
          "trends": {
            "$ref": "#/components/schemas/TrendAnalysis",
            "description": "Trend analysis data"
          },
          "summary": {
            "$ref": "#/components/schemas/app__models__analytics__ExecutiveSummary",
            "description": "Executive summary"
          }
        },
        "type": "object",
        "required": [
          "period",
          "generated_at",
          "project_metrics",
          "risk_metrics",
          "threat_metrics",
          "performance_metrics",
          "trends",
          "summary"
        ],
        "title": "DashboardMetricsResponse",
        "description": "Dashboard metrics response"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectAnalyticsResponse": {
        "properties": {
          "project_info": {
            "additionalProperties": true,
            "type": "object",
            "title": "Project Info",
            "description": "Project information"
          },
          "risk_assessment": {
            "additionalProperties": true,
            "type": "object",
            "title": "Risk Assessment",
            "description": "Risk assessment details"
          },
          "attack_paths": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Attack Paths",
            "description": "Identified attack paths"
          },
          "recommendations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Recommendations",
            "description": "Security recommendations"
          },
          "threat_intelligence": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threat Intelligence",
            "description": "Threat intelligence data"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          }
        },
        "type": "object",
        "required": [
          "project_info",
          "risk_assessment",
          "attack_paths",
          "recommendations"
        ],
        "title": "ProjectAnalyticsResponse",
        "description": "Detailed project analytics response"
      },
...
      "ProjectMetrics": {
        "properties": {
          "total_projects": {
            "type": "integer",
            "title": "Total Projects",
            "description": "Total number of projects"
          },
          "recent_projects": {
            "type": "integer",
            "title": "Recent Projects",
            "description": "Recently created projects"
          },
          "active_projects": {
            "type": "integer",
            "title": "Active Projects",
            "description": "Currently active projects"
          },
          "completion_rate": {
            "type": "number",
            "title": "Completion Rate",
            "description": "Project completion rate percentage"
          },
          "status_distribution": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Status Distribution",
            "description": "Distribution by project status"
          }
        },
        "type": "object",
        "required": [
          "total_projects",
          "recent_projects",
          "active_projects",
          "completion_rate",
          "status_distribution"
        ],
        "title": "ProjectMetrics",
        "description": "Project-related metrics"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}